feat: rename util func
